requirements.txt
streamlit
torch
torchvision
tensorflow-cpu==2.12.0
pillow
numpy